Move the interface of domains to it's own header file.
ai.{h,cpp} is already crowded with two different interfaces.  The
following patch set will make this much worse, thus we start by
refactoring.
