+++ src/analyses/Makefile
@@ -1,4 +1,5 @@
 SRC = ai.cpp \
+      ai_domain.cpp \
       call_graph.cpp \
       call_graph_helpers.cpp \
       constant_propagator.cpp \
